Resolve symlinks when loading app:/// and file:/// URLs on Windows (#11)
The WinRT file API does not resolve symlinks when used in Win32 apps.
This change replaces the WinRT file API with calls to C++ STL fstream
functions which does resolve symlinks on Win32 builds.
